Don't delete variables when doing tables
If you set a variable with flateSetVar() then it will only be visible
within a #BEGINTABLE zone once. That in itself is perhaps a side
effect.

However it can be useful to have access to variables that are not
necessarily part of the flateDumpTableLine() collection.

So let's not delete ##variables## from within flateDumpTableLine().

Now any variables that you set *before* calling flateDumpTableLine() will
be available to all calls of flateDumpTableLine().

I've checked with valgrind and this doesn't cause any memory leak. All
memory references are still being caught by flateFreeMem().

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>